Initial work on Events
- created declaration for PAL events
- move declaration from PAL header to to events header
- update CMakes to include these
- add implementation of bool timer event in ChibiOS
- enable bool timer event in interpreter
- adjust managed heap size for Discovery4 target
